move traits to own module
