+++ src/encryption/mod.rs
@@ -1,3 +1,5 @@
+mod traits;
 mod aes128_cbc;
 
+pub use traits::Encryption;
 pub use aes128_cbc::Aes128CbcEncryption;
